chore(scm): avoid dropping child before wait (#9564)
### Description

In the case of an error when parsing `git` output. We would drop a
`Child` without `wait`ing on it which results in a zombie process as the
pid is never reaped.

From [Rust
docs](https://doc.rust-lang.org/std/process/struct.Child.html#warning)

> On some systems, calling
[wait](https://doc.rust-lang.org/std/process/struct.Child.html#method.wait)
or similar is necessary for the OS to release resources. A process that
terminated but has not been waited on is still around as a “zombie”.
Leaving too many zombies around may exhaust global resources (for
example process IDs).

> The standard library does not automatically wait on child processes
(not even if the Child is dropped), it is up to the application
developer to do so. As a consequence, dropping Child handles without
waiting on them first is not recommended in long-running applications.

When there was a parse error we would `kill` the child process, but
never reap the pid. This PR ensures we make a best effort to do just
that. The way I'm calling wait is probably overkill, but I wanted to
ensure that we don't introduce any accidental waiting on a process that
didn't receive the kill signal.

Sources for comments:
 - [unix](https://man7.org/linux/man-pages/man2/kill.2.html)
-
[windows](https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-terminateprocess#return-value)

### Testing Instructions

I have done some manual confirmation that this works for a command like
`bash -c "sleep 100"` where it will

Hoping to get someone from
https://github.com/vercel/turborepo/issues/9455 to test this out in a
canary and confirm this helps.