fix(mfe): multiple versions (#9595)
### Description

We had a bug where if there were multiple MFE configs that contained the
same application we would defer to config name ordering for which config
would get selected. This PR makes it so we now select the newest
versioned configuration.

⚠️ Fix a bug I introduced in
https://github.com/vercel/turborepo/pull/9582 where I cast the relative
unix path to a basic `str` and I then misused it with `join_component`.

### Testing Instructions

Updated and added unit tests. Tested on a repo with multiple overlapping
configs with different versions.