ci(examples): test every package manager in Support Policy on `basic` (#9593)
### Description

Recently, I updated external dependencies for the `basic` example and
tested it with pnpm. Everything looked good - but our CI _only_ tests
`basic` for pnpm. I hadn't realized that I broke one of the other
package managers.

Instead of relying on making sure that we hand-test every package
manager, let's test each one in CI.

This PR also removes a few of the examples from the test suite since we
are not maintaining them. Because they are community maintained, we
won't rely on them for our CI.

### Testing Instructions

CI 🙏

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>