docs: mention module resolution differences for package managers (#9665)
### Description

We've gotten reports before about the documentation being wrong when
it's actually an issue with a package manager behaving in a bizarre way
or a specific setting for a package manager not working with the
strategy proposed in the documentation.

We won't be able to document every edge case or odd behavior of every
package manager's module resolution algorithm, so I'd prefer to make
mention of this in the docs.

We do our best. 😬

### Testing Instructions

👀