fix(cache): no longer let remoteCache.enable override force (#9684)
### Description

Fixes #9681

We were setting remote cache items to whatever the value of
`remoteCache.enable` was. This incorrectly re-enabled parts of the cache
that had previously been disabled. The fix is to only care if users
explicitly disable remote caching, this works since if
`remoteCache.enable` isn't specified it is assumed to be `true`.

### Testing Instructions

Added failing unit test in first commit, it passes after the second
commit.