fix(examples): basic generator (#9687)
### Description

Generators in the basic example broke as part of migrating to eslint 9
and converting the ui package to ESM. This fixes the generators by only
converting the eslint config file to mjs and not the entire package.

In the longterm we need to update the version of node-plop that is used
in turbo-gen, but that requires updating our internal packages as well
which will take a bit and require a major version change.

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->