chore(examples): mark core supported examples (#9707)
### Description

The community has been contributing more and more examples as Turborepo
usage has grown. This is _so exciting_ for those of us on the core team,
but we've had a hard time maintaining so many examples. It's a lot of
work to keep up with the breadth of tools that work great with
Turborepo, and the JavaScript ecosystem moves fast.

Because of this, we've come up with a strategy so that we can set
expectations correctly with the community. We're picking out ~5 examples
that we're committing to maintaining as shining representatives of
JavaScript monorepo greatness, leaving the rest to be maintained by the
community.

We will communicate this directly in `create-turbo` invocations, leaving
a note for users when they pick out a community-supported example.
 
### Testing Instructions
Should not show a `Note` log: `devcreateturbo --example
https://github.com/vercel/turborepo/tree/shew-0f97e/examples/basic`
Should show a `Note` log: `devcreateturbo --example
https://github.com/vercel/turborepo/tree/shew-0f97e/examples/with-docker`