chore(turbo): include schema in turbo package (#9714)
### Description

Add a new `@turbo/types#copy-schema` task which will copy the current
schema into `packages/turbo`. We do this since we explicitly exclude the
`turbo` package from our workspace so it's easier to have this hang off
of `@turbo/types`.

We then add this task to the `build` target which is a dependency of the
`publish-turbo` target which is what gets run on release:
https://github.com/vercel/turborepo/actions/runs/12796720464/job/35678219119#step:11:2

Resolves #7466

### Testing Instructions

Do a test run of the release and make sure we end up running the new
`copy-schema` task. [Started
here](https://github.com/vercel/turborepo/actions/runs/12812613189),
will take a sec to get to the point in the release where we hit this
codepath.

You can see schema gets copied into the `turbo` package
[here](https://github.com/vercel/turborepo/actions/runs/12812613189/job/35726384650#step:11:42)