ci(examples): update version in examples on release (#9823)
### Description

Updates the `turbo` version in examples whenever we make a release. This
ignores canaries so we don't ship those to unexpecting users.

I'm not sure that this is a sustainable solution as the community
continues to add more examples since the many `pnpm update` calls might
take awhile. However, I feel like this isn't a performance-critical path
for us, so going to go with it until someone tells me not to/thinks of a
more clever way. Note that whatever we go with needs to 1) bump the
version in package.json and 2) update the lockfile.

---------

Co-authored-by: Thomas Knickman <tom.knickman@vercel.com>