ci: add `check-types` task to examples testing (#9824)
### Description

In the interest of keeping the examples robust, let's also run the newly
added `check-types` task in CI to ensure those stay in the green when
people try them out for the first time, too.

Doing this uncovered some issues with existing examples, so glad to see
that this will get us to a more stable place!

A notable exception, unfortunately, is going to be the `check-types` in
`with-svelte`. Svelte appears to be doing some type generation that can
interrupt the type-checking process and make it flake out.

### Testing Instructions

CI