feat(bun): support `bun.lock` (#9783)
### Description

Closes #9628

Switch over from our usage of `yarn.lock` and directly support
`bun.lock` parsing.

We use `biome` to strip out the trailing commas that appear in the
`workspaces` and `packages` objects to leverage `serde` which I am more
competent in. We can switch to `biome` in the future and avoid a second
pass, but I do not know how to do correct deserialization logic for
`PackageEntry` (see `de.rs` for examples of what this looks like) in
`biome`.

I will call out is the fact that the lockfile keys in `bun.lock` aren't
used directly unlike other lockfile implementation. Instead we use
resolved package identifiers e.g. `package@(protocol:)?version`. The
keys themselves can shift when the underlying package does not change
resulting in incorrect behavior.

A quick example to illustrate:
 - `a` depends on `shared@1.0.0`
 - `b` depends on `shared@2.0.0` 
 - `shared@1.0.0` will get they key of `shared` since `a` is before `b` 
 - `shared@2.0.0` will get a key of `b/shared`
- If `a` updates to `shared@2.0.0`, then the `shared` key will now point
to `shared@2.0.0`
- If we used the key instead of the ident we would rebuild `b` since
it's key changed from `b/shared` to `shared` and not rebuild `a` since
it's key remained `shared`

This PR does not add support for `turbo prune`.

Reviewing each commit on it's own would probably be helpful.

### Testing Instructions

Unit tests. Manual testing on a `create-turbo` repository.