feat(boundaries): ignore svelte and vue files (but warn) (#9886)
### Description

SWC currently cannot parse svelte or vue. Therefore we ignore them when
it comes to boundaries, but we do emit a warning for users. Note that we
specifically do *not* emit a warning per file, since that could be a lot
of warnings. We do however emit per package, but that should hopefully
not be too excessive

closes #9874 

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->