docs(repo): add JSON Schema to `turbo.json` ref in `single-package-workspaces.mdx` (#9926)
### Description

<!--
  ✍️ Write a short summary of your work.
  If necessary, include relevant screenshots.
-->

Add a JSON schema to the `turbo.json` reference in
`single-package-workspaces.mdx`, found in the Guides section.

This is useful for newcomers who are setting up Turborepo and learning
how to set that file up from this guide - they will fall into the pit of
success and have the JSON schema set up.

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->

N/A