chore: standardize type check job name to check-types (#9973)
### Description

This pull request standardizes the type checking task name to
`check-types` across all example repos. This is good so that new users
can learn one way and feel confident they will be able to run this
across all repos that need it.

### Testing Instructions

Running `turbo check-types` in `with-shell-commands` example

![Screenshot 2025-02-14 at 12 34
21 PM](https://github.com/user-attachments/assets/5252ab8d-8d30-423a-abd5-39e5451ead24)

Running `pnpm check-types` in `with-typeorm` repo

![Screenshot 2025-02-14 at 12 34
33 PM](https://github.com/user-attachments/assets/b3ea332f-915d-472f-a36b-e7f5d7f76207)

Running `pnpm check-types` in `with-gatsby/apps/web`
![Screenshot 2025-02-14 at 12 43
06 PM](https://github.com/user-attachments/assets/c09a36f9-a6df-4e2f-adea-bc062ff2a181)


Running `pnpm check-types` and `pnpm build` in `with-vue-next/apps/web` 
![Screenshot 2025-02-14 at 12 44
23 PM](https://github.com/user-attachments/assets/953348d2-8e8b-4759-be3a-cdcc4d469bbd)

---------

Co-authored-by: Anthony Shew <anthony.shew@vercel.com>