docs: update TypeScript wildcard recommendations (#9983)
### Description

@thdxr kindly pointed out that the wildcard recommendations in our docs
weren't matching what he was seeing in his editor. I investigated and am
similarly seeing that wildcard recommendations work as one would hope.
The last time that I tried this, there was some limit at 10 exports
paths or 10 variable exports (my memory is fuzzy). But that appears to
be removed now in latest versions of TypeScript with modern settings
(which is what the Turborepo docs on other tools always assume).

Additional credit to @baptisteArno who also was pointing this out to me
in #9827. I'm going to use this PR since it has a few more spots in docs
where things are updated, and has additional updates to our examples.
But let the record show that @baptisteArno deserves credits here!

### Testing

I hand-tested this in my editor and here's what I see. One or more of
these didn't used to work but now it does.

![CleanShot 2025-02-15 at 23 32
38@2x](https://github.com/user-attachments/assets/ac7a44b3-8fe3-4a72-a580-2abb151d2100)

Additionally, here is my editor recommending 10+ entrypoints.
![CleanShot 2025-02-15 at 23 35
19@2x](https://github.com/user-attachments/assets/695087e4-5e29-474a-b573-cadd62e14fab)