feat(turborepo): --continue=independent-tasks-only
Convert `--continue` into a tri-state option, introducing a new mode that
allows independent tasks to continue running when an error occurs (dependent
tasks will be canceled).

The possible values for `--continue` are now:
- `none` (default) -- on error: cancel all tasks
- `independent-tasks-only` -- on error: cancel dependents, but continue others
- `all` -- on error: always continue, including dependents

For backwards compatibility, `false` (i.e. `none`) and `true` (i.e. `all`) are
still accepted, but will emit a deprecation warning. Setting `--continue`
without a value is equivalent to setting it to `all`