fix(tui): no longer error if preferences cannot be read (#10004)
### Description

Fixes #9999

The TUI should function even if we fail at loading preferences. 

### Testing Instructions

Add test to verify that if the preferences file contains invalid JSON a
loader is still created. User probably was hitting a perms issue or some
other FS issue and not an invalid JSON, but the test is just meant to
show we don't crash.