chore: extract logic for defaulting to run into helper (#10011)
### Description

Refactor to pull the logic for getting a `Command` from `cli_args` into
a named helper.

The `get_command` helper further uses a helper that makes it clear that
`Command::Run` is the default if no command is specified when Turborepo
is invoked.

There's still some future work to do based on the comment here:


https://github.com/vercel/turborepo/blob/b6a97bdc947f0b4855a1c9051df7eb0295cae39e/crates/turborepo-lib/src/cli/mod.rs#L1250

to make `cli_args` to be non-mutable.