chore: bump Svelte example to Svelte v5 (#9981)
### Description

This PR migrates the Svelte example to Svelte v5.

It was performed by running `pnpx npm-check-updates -u` in all packages
in `examples/with-svelte` directory.

Also ran `pnpx sv migrate svelte-5` in
`examples/with-svelte/packages/ui`.

### Testing Instructions

In `examples/with-svelte`, run `pnpm dev` and visit `web` and `docs`
sites to verify it's working.

This does not impact any automated tests.

---------

Co-authored-by: Anthony Shew <anthonyshew@gmail.com>