chore(turbo_json): remove exterior mutability from loader (#10066)
### Description

`TurboJsonLoader::load` only took a mutable self reference because we
were using `HashMap` to cache the results. This PR changes the data type
we use to cache results and removes the exterior mutability.

We move to use a `FixedMap` which has the following properties:
 - Only works with the keys provided at construction
 - Append only
 - Thread safe

I went with this just because the implementation is straightforward and
matches our use case since we know all of the possible places a
`turbo.json` will be before we start loading them.

Future work can be reworking boundaries code to no longer eagerly load
all of the `turbo.json`s at the start of execution.

### Testing Instructions

Existing test suite. Additional unit tests for `FixedMap`

---------

Co-authored-by: Nicholas Yang <nicholas.yang@vercel.com>