docs: fix JSDoc annotation for ESLint flat configs of basic example (#10089)
### Description

This is a very minor change I noticed when copying the basic example.
The JSDoc annotation for the ESLint flat configs is incorrect,
indicating a config object instead of an array of config objects, which
prevents the IDE from providing proper autocomplete.

### Testing Instructions

The IDE autocomplete feature allows to very quickly validate the change.