fix(prune): support copying turbo.jsonc (#10107)
### Description

Realized we weren't considering `jsonc` as a possible source when
copying files for `prune`.

Some future work is to make `prune` work with the now exposed
`root_turbo_json()` method off of the configuration object, but didn't
have time for that refactor at the moment.

### Testing Instructions

Added `turbo.jsonc` test to the prune e2e tests