feat(boundaries): implicit dependencies (#10117)
### Description

Allow users to allowlist implicit dependencies for boundaries. This is
useful in cases where testing libraries or frameworks can inject globals
such as `server-only`

Can be reviewed commit by commit
### Testing Instructions

Added test to boundaries. Note the snapshots do *not* change since no
errors are produced. To verify stuff is working as intended, try
removing the `implicitDependencies` key and note that an error is
produced.