fix(bun): properly handle bun lockfile keys (#10137)
### Description

We were not properly handling recursing when traversing the lockfile.
Resolving a package would give a `package@version` key instead of the
key into the `packages` array. Our `all_dependencies` implementation
assumed that we were given the key into `packages`.

This is a little odd as the keys can have different "paths" (keys to the
packages object) that map to identical entries. This PR adds a check to
verify that the entries are in fact the same package by comparing
checksums.

### Testing Instructions

Fixed unit tests and added one with mismatched checksums.
`create-turbo` using bun should no longer display ` WARNING Unable to
calculate transitive closures: No lockfile entry found for
'@types/react@19.0.10'`