docs: redirect `/repo/docs` to `/docs` (#10232)
### Description

Replaces the `turbo.build/repo/docs` path to `turbo.build/docs` now that
we have moved the documentation for Turbopack to a different place.

### Testing Instructions

A bit tough to review, I apologize. I can break this up into smaller PRs
if we wish (since redirects will be in place for the old path pattern).