feat(tui): configurable scrollback length (#10247)
### Description

Allow users to set a different value for scrollback length. I'm also
taking a liberty to double the scrollback length's default value here
(cc @chris-olszewski to tell me if this isn't a good idea)

#### Original asks:
- https://github.com/vercel/turborepo/discussions/10234
-
https://github.com/vercel/turborepo/issues/8643#issuecomment-2749429985

> [!NOTE]
> If you're wondering why we don't dynamically set this to whatever the
terminal's scrollback is, terminals don't expose this value. The closest
thing to a spec for terminals is [the xterm
spec](https://invisible-island.net/xterm/ctlseqs/ctlseqs.txt), so we
don't know what we can depend on.

### Questions

Should this also be:
- A `turbo.json` config?
- Personally, I don't think so. Our reasoning for configurations in
`turbo.json` tends to be that it should be something that makes sense
for *anyone* in the repo, and setting some non-default scrollback length
doesn't sound desirable.
- A flag?
- This sounds reasonable, but omitting here unless an ask comes in for
it.

### Testing Instructions

Try it!