docs: list pnpm as first package manager (#10293)
### Description

After spending enough time with monorepo users, we've developed the
opinion that pnpm should be the preferred package manager that we point
people to.

We're nudging pnpm to the top of the tabs where we list package managers
to lean on this preference.

> [!NOTE]
> While working on this, I found out that these tabs aren't syncing with
each other. Changing to pnpm only changes for the one you clicked, not
all `<PackageManagerTabs />`. The bug is not introduced by this PR, and
I'll fix it separately.

### Testing Instructions

👀