fix(codemod): check for pipeline key before using it (#10295)
### Description

After users successfully execute codemods for v1, their package manager
installation at the end can fail, as seen in #8495. This means that the
`pipeline` key would disappear, but the `turbo` in the repository might
be stuck at v1 still.

If they try to run the codemods again, the codemod will try to execute
codemods for v1 again, since that's what will be found in package.json.
At this point, it will error, when it can't find the `pipeline` key.

This PR introduces a guard that we can use to early exit if the
`pipeline` key doesn't exist.

### Testing Instructions

I'm reproducing #8495 with these steps:
1. `npx create-turbo@latest`
2. Bump `turbo` in package.json down to `1.0.0`.
3. Error!

With the changes in this PR, the codemods are effectively skipped, and
we make it past them succesfully.