feat: allow specifying concurrency in config / environment (#10236)
### Description

This adds a couple new ways to configure task run concurrency: 

- A `TURBO_CONCURRENCY` environment variable
- A `concurrency` configuration value in `turbo.json`

These options should follow the same precedence rules for other options
which are made available in multiple locations. I think the environment
variable is the most valuable to me (IMO), as this would make it a lot
easier to turn off or drastically reduce concurrency in CI scenarios
with restricted performance, where running multiple test suites at the
same time could cause flaky results due to timeouts, or too much
resource contention in general scenarios. Being able to specify this in
the configuration file is maybe less valuable, but there are situations
where it might be desirable.

The `--concurrency` option is obviously already available and certainly
usable in combination with other changes (enforcing the use of an alias,
or setting up different `package.json` scripts for CI and local
development), but enabling this to be enforced via the config or an
environment variable makes it a lot easier to turn off concurrency in
CI, for example.

There is some active desire for this feature to exist:
- https://github.com/vercel/turborepo/discussions/7263
- https://github.com/vercel/turborepo/discussions/7493

Notably, this does _not_ enable configuring concurrency per task, which
was part of one of the discussions here. That seemed like a much larger
lift vs. just adding an environment variable and config option that
behaved exactly like the CLI argument already does. There is still some
demand for this "simple" option.

### Testing Instructions

1. Build `turbo`
2. Check that you can now specify a top-level "concurrency" property in
your `turbo.json`, with the same behaviour as the CLI argument.
3. Check that you can now specify a `TURBO_CONCURRENCY` environment
variable, with the same behaviour as the CLI argument.
4. Check that this settings maintains expected preference (i.e. CLI arg
> env var > turbo.json)

I've done this checks (with some print debugging) on repos where I use
`turbo` and it seems to work as expected. I've added some unit and
integration tests where it seemed to make the most sense, but there were
some spots where it was unclear if I should add something. If I've
missed a spot or need some additional checks, let me know!