chore: use Turbopack for documentation builds (#10312)
The fumadocs MDX transformation seems to be off with Turbopack, and some
MDX frontmatter appears to be missing.

There is unfortunately no way to satisfy these `Package shiki can't be
external` warnings, because there are two packages each requesting
different versions of shiki. The only solution would be to somehow use a
combination of `fumadocs-core` and `@shikijs/rehype` versions that both
use the same version of `shiki`.

```
fumadocs-core/dist
The package resolves to a different version when requested from the project directory (3.2.1) compared to the package requested from the importing module (2.5.0).

@shikijs/rehype/dist
The package resolves to a different version when requested from the project directory (2.5.0) compared to the package requested from the importing module (3.2.1).
```